Try to figure out why encoding/binary is so bloody slow.
bx is a test that extracts the relevant parts of encoding/binary
and does the same thing here.

The performance is very close to 'bi', other than I skipped an
indirection through endianness so that could explain the 20% gain.

It immediately becomes clear what's going on. We not only do
reflections on the data to see what type it is, we also do a
reflection on each element in the slice just in case the slice
changes type in the middle. This is beyond wasteful.